ok reorganised the code to make it a little easier to follow
