+++ src/models/AbstractStyledComponent.js
@@ -3,7 +3,7 @@ import { Component } from 'react'
 import PropTypes from 'prop-types'
 
 import { CHANNEL } from './ThemeProvider'
-import { StyleSheet } from './BrowserStyleSheet'
+import StyleSheet from './StyleSheet'
 
 export default class AbstractStyledComponent extends Component {
   static isPrototypeOf: Function
